refactor(discord): group server Discord utils under discord/
Move cache, oauth, and guild helpers into server/utils/discord/ to match
the audit/ folder pattern and simplify imports.
