+++ test/unit/server/utils/discord/get-current-user.spec.ts
